atividades/08-07-2024 - gRPC/client/pyproject.toml
[tool.poetry]
name = "client"
version = "0.1.0"
description = ""
authors = ["TheLastAurora <vitoribeiro0024@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
grpcio-tools = "^1.64.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
